requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==3.0.0
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
Pillow==10.2.0
numpy==1.26.4
matplotlib==3.8.2
scikit-learn==1.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
stripe==7.8.0
